Remove casting

+++ ps3/src/simplifynodes.c
@@ -27,7 +27,7 @@ Node_t* simplify_types ( Node_t *root, int depth )
 	
 	int index_children = root->children[1]->n_children;
 	root->data_type.n_dimensions = index_children;
-	root->data_type.dimensions = (int)malloc(sizeof(int)*index_children);
+	root->data_type.dimensions = malloc(sizeof(int)*index_children);
 	for (int j = 0; j < index_children; j++){
 		root->data_type.dimensions[j] = root->children[1]->children[j]->int_const;
 		//node_finalize(root->children[1]->children[j]);
